Move methods from view to bloc class
